backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
psycopg2-binary==2.9.9